moved data from index.js to App.js componentDidMount
