+++ src/App.js
@@ -1,4 +1,4 @@
-import {useState} from "react";
+import {useState, useEffect} from "react";
 import ActionsMenu from "./Components/ActionsMenu";
 import AccountInfo from "./Components/AccountInfo";
 import Expenses from "./Components/Expenses";
@@ -194,6 +194,17 @@ function App() {
     let obj = {users: [...users]}
     localStorage.setItem("bankdata", JSON.stringify(obj))
   }
+
+  useEffect(()=> {
+    let test = localStorage.getItem("bankdata");
+    let obj = {users: [
+      {id: 1, username: "JohnD", firstName: "John", lastName: "Doe", email: "john@gmail.com", password: "1234", balance: 500, expenses: [{expenseName: "Electricity bill", amount: 500, expenseID: 1}]},
+      {id: 2, username: "JaneD", firstName: "Jane", lastName: "Doe", email: "jane@gmail.com", password: "4321", balance: 1000, expenses: []}
+    ]}
+    if(test == null){
+      localStorage.setItem('bankdata', JSON.stringify(obj))
+    }
+  }, [])
 
   return (
     <Router>
